Add configure script
Now is possible to do the typical:
    % ./configure
    % make
    % make install

Configure script is not so smart as the one from autotools but it should
get the job done. It is also interactive - the reason behind this was to
avoid unwanted surprises for the user - script forces him to acknowledge
the values or at least gives him the nudge to look into Makefile.config.

Signed-off-by: Petr Ospalý <petr@ospalax.cz>