Change curses to ncurses
I am still not sure which of them is the more portable name, but my
focus is on ncurses and curses (lib or header file) is a symlink to
ncurses on most installations so I will be consistent and stick with
the ncurses naming.

Although I am quite confident that on the most systems these two are
interchangeable (ncurses is 'New curses' anyway).

Signed-off-by: Petr Ospalý <petr@ospalax.cz>